Css gradient Colors

+++ server/controllers/EmployeeControllers.js
@@ -1,31 +1,39 @@
-import Employee from "../models/EmployeeModel.js";
+import mongoose from 'mongoose'
+import Employee from '../models/EmployeeModel.js'
 
 const AddEmployee = async (req, res, next) => {
   try {
-    const { name, address, experience, phone, email, endDate, idCard } =
-      req.body;
+    const {
+      name,
+      address,
+      experience,
+      phone,
+      email,
+      endDate,
+      idCard,
+    } = req.body
     if (!name || !address || !phone || !email || !idCard) {
       return res.json({
-        message: "Please fill all fields",
-      });
+        message: 'Please fill all fields',
+      })
     }
     if (phone.length < 11) {
       return res.json({
-        message: "Phone Number contains 11 digits",
-      });
+        message: 'Phone Number contains 11 digits',
+      })
     }
     if (idCard.length < 13) {
       return res.json({
-        message: "NIC contains 13 digits",
-      });
+        message: 'NIC contains 13 digits',
+      })
     }
-    const exist = await Employee.findOne({ email });
+    const exist = await Employee.findOne({ email })
     if (exist) {
-      return res.json({ message: "Email already exists" });
+      return res.json({ message: 'Email already exists' })
     }
-    const NICexist = await Employee.findOne({ idCard });
+    const NICexist = await Employee.findOne({ idCard })
     if (NICexist) {
-      return res.json({ message: "NIC already exists" });
+      return res.json({ message: 'NIC already exists' })
     }
 
     const newEmployee = new Employee({
@@ -36,38 +44,41 @@ const AddEmployee = async (req, res, next) => {
       email,
       endDate,
       idCard,
-    });
-    await newEmployee.save();
-    return res.json({ message: "Employee added successfully" });
+    })
+    await newEmployee.save()
+    return res.json({ message: 'Employee added successfully' })
   } catch (error) {
     // next(error);
-    next(error);
+    next(error)
   }
-};
+}
 
 const getALLEmployees = async (req, res) => {
   try {
-    const employees = await Employee.find();
-    return res.json(employees);
+    const employees = await Employee.find()
+    return res.json(employees)
   } catch (error) {
-    console.log(error);
+    console.log(error)
   }
-};
+}
 
 const getEmployee = async (req, res) => {
   try {
-    const employee = await Employee.findById(req.params.id);
-    return res.json(employee);
+    const employee = await Employee.findById(req.params.id)
+    return res.json(employee)
   } catch (error) {
-    console.log(error);
+    console.log(error)
   }
-};
+}
 
 const updateEmployee = async (req, res) => {
-  const { name, address, experience, phone, email, endDate, idCard } = req.body;
+  const { name, address, experience, phone, email, endDate, idCard } = req.body
 
+  if (!name || !address || !phone || !email || !idCard) {
+    return res.json({ message: `Please fill the required fields` })
+  }
   try {
-    const updatedEmployee = await Employee.findByIdAndUpdate(req.params.id, {
+    await Employee.findByIdAndUpdate(req.params.id, {
       name,
       address,
       experience,
@@ -75,20 +86,25 @@ const updateEmployee = async (req, res) => {
       email,
       endDate,
       idCard,
-    });
-    res.status(200).json(`Employee updated successfully`);
+    })
+    return res.status(200).json({ message: `Employee updated successfully` })
   } catch (error) {
-    console.log(error);
+    next(error)
   }
-};
+}
 const deleteEmployee = async (req, res) => {
   try {
-    const daletedEmployee = await Employee.findByIdAndDelete(req.params.id);
-    return res.json(daletedEmployee);
+    console.log(req.params.id)
+    if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+      return res.json({ message: 'id is not valid' })
+    }
+
+    await Employee.findByIdAndDelete(req.params.id)
+    return res.json({ message: `Employee deleted successfully` })
   } catch (error) {
-    console.log(error);
+    next(error)
   }
-};
+}
 
 export {
   getALLEmployees,
@@ -96,4 +112,4 @@ export {
   getEmployee,
   updateEmployee,
   deleteEmployee,
-};
+}
